feat(syscall): update syscall argument types to improve consistency and safety

+++ kernel/src/syscall/linux/path.rs
@@ -4,9 +4,9 @@ pub(super) fn linux_translate_user_path(path_ptr: u64) -> Option<String> {
     super::handlers::helpers::read_user_str_safe(path_ptr).map(linux_translate_path)
 }
 
-pub(super) fn linux_translate_at_path(dirfd: u64, path_ptr: u64) -> Result<String, i32> {
+pub(super) fn linux_translate_at_path(dirfd: i32, path_ptr: u64) -> Result<String, i32> {
     let path = super::handlers::helpers::read_user_str_safe(path_ptr).ok_or(EFAULT)?;
-    if path.starts_with('/') || (dirfd as i32) == LINUX_AT_FDCWD {
+    if path.starts_with('/') || dirfd == LINUX_AT_FDCWD {
         return Ok(linux_translate_path(path));
     }
     let base = linux_fd_path(dirfd as u32)?;
